+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 5.6.2 (Unreleased)
+
+### Features
+
+ - Make install of WiringPi optional ([#412](https://github.com/kizniche/mycodo/issues/412))
+
+
 ## 5.6.1 (2018-02-27)
 
 ### Features